hitbox collider now flips and destroys enemies
Need to work on making it feel better like more responsive and optimize the collider distance. I kept getting killed when I was trying to attack which just shouldn't happen. I also noticed a bug where when you attack and move back and forth it hecks up the animation. I would also like to be able to attack while in the air